pyproject.toml
[tool.poetry]
name = "metrica"
version = "0.1.0"
description = ""
authors = ["egor187 <py.egor.py@gmail.com>"]
packages = [
    { include = "Metrica_project" },
    { include = "games" },
    { include = "users" },
    { include = "telegram_bot" },
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.6"
python-dotenv = "^0.15.0"
django-bootstrap4 = "^2.3.1"
python-telegram-bot = "^13.4.1"
rollbar = "^0.15.2"
django-anymail = {extras = ["mailgun"], version = "^8.2"}
whitenoise = "^5.2.0"
django-debug-toolbar = "^3.2"
Pillow = "^8.1.2"
django-filter = "^2.4.0"
Faker = "^6.6.3"
PyJWT = "^2.0.1"
django-summernote = "^0.8.11"
django-redis = "^4.12.1"
django-swagger-render = "^0.1.1"
pytest = "^6.2.3"
celery = "^5.0.5"
boto3 = "^1.17.89"
django-storages = "^1.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
